meny funkar med att ladda bana fran fil, player smalare

+++ tddd23blokz/src/com/tddd23/blokz/Constants.java
@@ -7,7 +7,7 @@ import com.badlogic.gdx.math.Vector2;
  */
 public class Constants {
 	public static final float MAX_FPS = 60;
-	public static final float SIZE = 16f;
+	public static final int SIZE = 16;
 	public static final float MAX_FALLING_SPEED = 5f*MAX_FPS ;
 	public static final float JUMPING_SPEED = 4.1f*MAX_FPS ;
 	public static final float MAX_MOVING_SPEED = 100f*MAX_FPS ;
